pyproject.toml
[tool.poetry]
name = "streamlit-pr"
version = "0.1.0"
description = ""
authors = ["dongho <dons97@naver.com>"]
readme = "README.md"
packages = [{include = "streamlit_pr"}]

[tool.poetry.dependencies]
python = "3.10.0"
transformers = "^4.28.1"
pytorch-lightning = "^2.0.2"
pyaml = "^21.10.1"
streamlit = "^1.21.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
